cleaning up and modularisation
